Update data checks to use if
All try except blocks that were attempting to check data now use the
equivalent if statement and manually raise the appropriate error instead of
using the `six` library. The `six` library has been removed from the list of
required packages as well. There was a single try except block that did not
raise an error for the search-members endpoint that was used to populate one
of the local variables that has also been updated.